remove master hardcoding number
