+++ pkg/config/data/etcd.conf.tpl
@@ -1,0 +1,19 @@
+ETCD_ADVERTISE_CLIENT_URLS={{ .AdertiseUrls }}
+ETCD_CERT_FILE={{ .AdertiseUrls }}
+ETCD_CLIENT_CERT_AUTH={{ .AdertiseUrls }}
+ETCD_DATA_DIR={{ .AdertiseUrls }}
+ETCD_ELECTION_TIMEOUT={{ .AdertiseUrls }}
+ETCD_HEARTBEAT_INTERVAL={{ .AdertiseUrls }}
+ETCD_INITIAL_ADVERTISE_PEER_URLS={{ .AdertiseUrls }}
+ETCD_INITIAL_CLUSTER={{ .AdertiseUrls }}
+#ETCD_INITIAL_CLUSTER=master-000000={{ .AdertiseUrls }}
+ETCD_KEY_FILE={{ .AdertiseUrls }}
+ETCD_LISTEN_CLIENT_URLS={{ .AdertiseUrls }}
+ETCD_LISTEN_PEER_URLS={{ .AdertiseUrls }}
+ETCD_NAME={{ .AdertiseUrls }}
+ETCD_PEER_CERT_FILE={{ .AdertiseUrls }}
+ETCD_PEER_CLIENT_CERT_AUTH={{ .AdertiseUrls }}
+ETCD_PEER_KEY_FILE={{ .AdertiseUrls }}
+ETCD_PEER_TRUSTED_CA_FILE={{ .AdertiseUrls }}
+ETCD_QUOTA_BACKEND_BYTES={{ .AdertiseUrls }}
+ETCD_TRUSTED_CA_FILE={{ .AdertiseUrls }}
